+++ app/src/main/java/com/github/saikcaskey/pokertracker/presentation/main/DefaultDashboardFeaturePlannerComponent.kt
@@ -3,19 +3,19 @@ package com.github.saikcaskey.pokertracker.presentation.main
 import com.arkivanov.decompose.ComponentContext
 import com.github.saikcaskey.pokertracker.domain.CoroutineDispatchers
 import com.github.saikcaskey.pokertracker.domain.repository.EventRepository
-import com.github.saikcaskey.pokertracker.domain.components.MainPagerPagePlannerComponent
-import com.github.saikcaskey.pokertracker.domain.components.MainPagerPagePlannerComponent.*
+import com.github.saikcaskey.pokertracker.domain.components.DashboardFeaturePlannerComponent
+import com.github.saikcaskey.pokertracker.domain.components.DashboardFeaturePlannerComponent.*
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.flow.SharingStarted.Companion.Eagerly
 import kotlinx.datetime.LocalDate
 
-class DefaultMainPagerPagePlannerComponent(
+class DefaultDashboardFeaturePlannerComponent(
     componentContext: ComponentContext,
     private val onCalendarDayClicked: (LocalDate, Boolean) -> Unit,
     eventsRepository: EventRepository,
     dispatchers: CoroutineDispatchers,
-) : MainPagerPagePlannerComponent, ComponentContext by componentContext {
+) : DashboardFeaturePlannerComponent, ComponentContext by componentContext {
 
     private val coroutineScope = CoroutineScope(dispatchers.io)
