+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/components/DashboardFeatureComponent.kt
@@ -1,3 +1,3 @@
 package com.github.saikcaskey.pokertracker.domain.components
 
-interface MainPagerPageComponent
+interface DashboardFeatureComponent
