+++ libs/domain/src/main/java/com/github/saikcaskey/pokertracker/domain/components/DashboardFeaturePlannerComponent.kt
@@ -3,7 +3,7 @@ package com.github.saikcaskey.pokertracker.domain.components
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.datetime.LocalDate
 
-interface MainPagerPagePlannerComponent : MainPagerPageComponent {
+interface DashboardFeaturePlannerComponent : DashboardFeatureComponent {
     val uiState: StateFlow<UiState>
 
     fun onShowDayDetail(day: LocalDate, hasEvent: Boolean)
