+++ libs/ui-compose/src/main/java/com/github/saikcaskey/pokertracker/ui_compose/components/main/MainContent.kt
@@ -7,8 +7,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.arkivanov.decompose.extensions.compose.pages.*
 import com.github.saikcaskey.pokertracker.domain.components.MainComponent
-import com.github.saikcaskey.pokertracker.domain.components.MainPagerPageDashboardComponent
-import com.github.saikcaskey.pokertracker.domain.components.MainPagerPagePlannerComponent
+import com.github.saikcaskey.pokertracker.domain.components.DashboardFeatureComponentImpl
+import com.github.saikcaskey.pokertracker.domain.components.DashboardFeaturePlannerComponent
 
 @Composable
 internal fun MainContent(component: MainComponent, modifier: Modifier = Modifier) {
@@ -37,8 +37,8 @@ internal fun MainContent(component: MainComponent, modifier: Modifier = Modifier
                 scrollAnimation = PagesScrollAnimation.Default,
             ) { idx, pageComponent ->
                 when (pageComponent) {
-                    is MainPagerPagePlannerComponent -> MainPagerPlannerContent(pageComponent)
-                    is MainPagerPageDashboardComponent -> MainPagerDashboardContent(pageComponent)
+                    is DashboardFeaturePlannerComponent -> MainPagerPlannerContent(pageComponent)
+                    is DashboardFeatureComponentImpl -> MainPagerDashboardContent(pageComponent)
                 }
             }
         }
